Code to detect rpm by listening to pulses in mic
Firstly, we need to get permissions from the user to access his/her
microphone. Then, when he starts to cycle, the cadence/rpm sensor sends
a "pulsing" sound to the microphone. We calculate RMS of this sound and
if it is greater than some defined threshold, we know that the spin bike
sent something. But before we can say that it was a **new** pulse, we
first need to make sure that it wasn't the old pulse, but still being
heard in the microphone. For this, we simply remember the timestamp of
the last pulse, and if the timestamp of the new pulse substracted by the
timestamp of the last pulse is greater than some defined threshold, we
assume we have a new pulse.

We still need to add code to emit this "on('pulse')" event because as
for now, this code won't emit that event.