Fix imports by removing * wildcard
If you use import * as Something from 'something', Something will be an
object. But if the module 'something' exports a function or a class
(constructor function), you can't simply call Something() then. That's
why you need to import it as import Something from 'something' so then
you can call it as Something().